el variable updates
make el in main event loop a ktl keyword object. Set it up to be
monitored so that it is always correct.

made “el” into “sunel” (so easier to find)

all numerical compares now wrapped with float (as default is a string
value, and the binary in radians ffs)

added extra check in opening to make sure waitfor does not wait during
sunrise for next sunset.